docker/app/requirements.txt
fastapi
psycopg2-binary
uvicorn[standard]
email-validator